Refactoring orchestrator.py
